Add code for exercises

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,17 +1,21 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-	const blogs = await Blog.find({});
-
+	const blogs = await Blog.find({}).populate("user", {
+		username: 1,
+		name: 1,
+	});
 	response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
 	const body = request.body;
+	const user = request.user;
 
-	if (!body.title || !body.url) {
-		return response.status(400).json({ error: "title or url missing" });
+	if (!user) {
+		return response.status(401).json({ error: "token invalid" });
 	}
 
 	const blog = new Blog({
@@ -19,16 +23,31 @@ blogsRouter.post("/", async (request, response) => {
 		author: body.author,
 		url: body.url,
 		likes: body.likes || 0,
+		user: user._id,
 	});
 
 	const savedBlog = await blog.save();
+	user.blogs = user.blogs.concat(savedBlog._id);
+	await user.save();
 
 	response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndDelete(request.params.id);
+blogsRouter.delete("/:id", middleware.userExtractor, async (request, response) => {
+	const user = request.user;
+	const blog = await Blog.findById(request.params.id);
 
+	if (!blog) {
+		return response.status(404).end();
+	}
+
+	if (blog.user.toString() !== user.id.toString()) {
+		return response
+			.status(401)
+			.json({ error: "only the creator can delete this blog" });
+	}
+
+	await Blog.findByIdAndDelete(request.params.id);
 	response.status(204).end();
 });
